started over to incorporate more components
