+++ src/App.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+
+
+class App extends Component {
+  render(){
+    return(
+      <h1> Hello World </h1>
+    )
+  }
+}
+
+
+export default App
