--- src/seeds.js
@@ -1,7 +1,0 @@
-let seeds = [
-  {title:  "Feed the Dog", content: "Feed the dog at 8PM. Two cans of dog food", completion: true},
-  {title: "Pick up Brother from the airport.", content: "Pick up brother at 9PM this Friday", completion: false},
-  {title: "Clean the Kitchen", content: "Sweep and mop floors.", completion: false}
-]
-
-export default seeds
